refactor(#3277): decompose ensureNativeStringHelpers — extract rope/flatten/UTF-8 core + concat/compare/slice (slice 2)
Stacked on #3275. Lift the remaining head/middle core builders (originally
lines 301-2284) verbatim out of ensureNativeStringHelpers into two NEW
cohesive sibling modules — this EMPTIES the god-function down to a ~48-line
driver (finalize guard + 4 type-idx consts + shared-bag build + 13 builder
calls + emitNativeHtmlWrapperHelpers):

- native-strings-core.ts   — __str_copy_tree/__str_utf8_to_flat/__str_flatten
                             (emitStrFlattenHelpers) + __str_to_utf8
                             (emitStrToUtf8Helper); flattenConsBody (#1588)
                             moved here (used only by flatten)
- native-strings-basics.ts — __str_concat/__str_buf_next_cap
                             (emitStrConcatHelpers), __str_equals/__str_compare
                             (emitStrCompareHelpers), __str_substring/charAt/
                             charAt_cp/slice/substr (emitStrSliceCharHelpers)

Inline strRef/flatStrRef/strDataRef/getFlattenIdx/wrapBodyWithFlatten locals
removed; makeNativeStrShared (from #3275) reconstructs them. Pure mechanical
relocation — byte-identical emit (prove-emit-identity IDENTICAL, 39/39).

native-strings.ts: 4,062 -> 2,025 LOC. ensureNativeStringHelpers now fully
decomposed into -shared/-core/-basics/-search/-transform/-rewrite.

✓
